requirements.txt
# Mira Voice Studio - Requirements
# Phase 1: Core Pipeline

# === PyTorch (Apple Silicon) ===
torch>=2.0.0
torchaudio>=2.0.0

# === TTS Engines ===
TTS>=0.22.0                     # Coqui TTS for Phase 1

# === Speech Processing ===
openai-whisper>=20231117
pydub>=0.25.1
soundfile>=0.12.1

# === Text Processing ===
nltk>=3.8.1

# === Audio Recording (Phase 4) ===
sounddevice>=0.4.6
numpy>=1.24.0

# === UI (Phase 3+) ===
gradio>=4.0.0

# === Utilities ===
tqdm>=4.66.0

# === macOS specific ===
pyobjc-framework-Cocoa>=9.0; sys_platform == "darwin"
pyobjc-framework-AVFoundation>=9.0; sys_platform == "darwin"
